+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Fix an issue that changes of insertion formats in the Snippets settings pane were not saved.
 - Fix an issue that the initial visibility state of the navigation bar was reversed.
+- [trivial] Fix an issue that the action name of the “Shift Left” action is not noted to the undo history.